fix(client): keep the status and body when the error envelope is unrecognized
On a non-2xx response the body was decoded into APIError unconditionally.
Because the codec ignores unknown keys, a JSON error of any other shape
(for example {"error":"upstream unavailable"}) decoded into a zero value
and was reported as "API error 0: ", discarding both the status code and
the server's message. A typed APIError is now returned only when the body
actually carries a code or a message; anything else keeps the previous
"status %d: %s" form.

- OrderStatusFilled.Oid widened from int to int64, matching
  OrderStatusResting.Oid: live order ids already exceed int32 (e.g.
  41569108691), which would overflow on a 32-bit build.
- CancelOrderResponse.Statuses gets the explicit json:"statuses" tag its
  OrderResponse sibling already has, and the stale comment claiming the
  codec matches keys exactly is corrected — sonic.ConfigStd is
  case-insensitive, like encoding/json.
- The remaining stdlib JSON calls (client.go json.Valid, api.go's two
  json.Marshal) go through jsonCodec; json.RawMessage stays a type usage.
- json.go documents that CopyString is load-bearing rather than a tunable:
  the websocket receive path hands decoders a string that points into a
  pooled buffer.
- json_wire_test.go closes the wire-format gaps: statuses decode for both
  response types, OrderWire with and without a cloid, trigger orders,
  cancel/cancelByCloid/batchModify action bytes, the APIError envelope and
  the client.post error paths. Expected bytes are taken from the recorded
  cassettes in testdata where one exists, otherwise from the struct tags.

+++ exchange_orders.go
@@ -29,7 +29,9 @@ type OrderStatusResting struct {
 type OrderStatusFilled struct {
 	TotalSz string `json:"totalSz"`
 	AvgPx   string `json:"avgPx"`
-	Oid     int    `json:"oid"`
+	// int64, matching OrderStatusResting.Oid: live order ids already exceed
+	// int32 (e.g. 41569108691), which would overflow a 32-bit int.
+	Oid int64 `json:"oid"`
 }
 
 type OrderStatus struct {
@@ -44,8 +46,9 @@ func (s *OrderStatus) String() string {
 }
 
 type OrderResponse struct {
-	// Explicit tag required: easyjson matches keys exactly (no stdlib-style
-	// case-insensitive fallback). The API sends "statuses".
+	// Explicit tag: the API sends "statuses". The codec would also match the
+	// field name case-insensitively (like encoding/json), but tagging keeps
+	// the wire key explicit and the intent documented.
 	Statuses []OrderStatus `json:"statuses"`
 }
